pyproject.toml
[tool.poetry]
name = "hbsir"
version = "0.1.0"
description = ""
authors = ["TerseTears <tersetears@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
pathlib = "^1.0.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
pyodbc = "^4.0.35"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
